requirements/base.txt
Django==1.11
django-environ==0.4.3
django-braces==1.11.0
django-crispy-forms==1.6.1
django-admin-bootstrapped==2.5.7
django-authtools==1.5.0
easy-thumbnails==2.4.1
djangorestframework==3.6.2
markdown==2.6.8
django-filter==1.0.2
openpyxl==2.4.7
reportlab==3.4.0
pyyaml==3.12